Refactored out code to separate modules.
