+++ src/chaining-functions.js
@@ -1,0 +1,92 @@
+export function nop() { }
+export function mirror(val) { return val; }
+export function pureFunctionChain(f1, f2) {
+    // Enables chained events that takes ONE argument and returns it to the next function in chain.
+    // This pattern is used in the hook("reading") event.
+    if (f1 === mirror) return f2;
+    return function (val) {
+        return f2(f1(val));
+    };
+}
+
+export function callBoth(on1, on2) {
+    return function () {
+        on1.apply(this, arguments);
+        on2.apply(this, arguments);
+    };
+}
+
+export function hookCreatingChain(f1, f2) {
+    // Enables chained events that takes several arguments and may modify first argument by making a modification and then returning the same instance.
+    // This pattern is used in the hook("creating") event.
+    if (f1 === nop) return f2;
+    return function () {
+        var res = f1.apply(this, arguments);
+        if (res !== undefined) arguments[0] = res;
+        var onsuccess = this.onsuccess, // In case event listener has set this.onsuccess
+            onerror = this.onerror;     // In case event listener has set this.onerror
+        this.onsuccess = null;
+        this.onerror = null;
+        var res2 = f2.apply(this, arguments);
+        if (onsuccess) this.onsuccess = this.onsuccess ? callBoth(onsuccess, this.onsuccess) : onsuccess;
+        if (onerror) this.onerror = this.onerror ? callBoth(onerror, this.onerror) : onerror;
+        return res2 !== undefined ? res2 : res;
+    };
+}
+
+export function hookUpdatingChain(f1, f2) {
+    if (f1 === nop) return f2;
+    return function () {
+        var res = f1.apply(this, arguments);
+        if (res !== undefined) extend(arguments[0], res); // If f1 returns new modifications, extend caller's modifications with the result before calling next in chain.
+        var onsuccess = this.onsuccess, // In case event listener has set this.onsuccess
+            onerror = this.onerror;     // In case event listener has set this.onerror
+        this.onsuccess = null;
+        this.onerror = null;
+        var res2 = f2.apply(this, arguments);
+        if (onsuccess) this.onsuccess = this.onsuccess ? callBoth(onsuccess, this.onsuccess) : onsuccess;
+        if (onerror) this.onerror = this.onerror ? callBoth(onerror, this.onerror) : onerror;
+        return res === undefined ?
+            (res2 === undefined ? undefined : res2) :
+            (res2 === undefined ? res : extend(res, res2));
+    };
+}
+
+export function stoppableEventChain(f1, f2) {
+    // Enables chained events that may return false to stop the event chain.
+    if (f1 === nop) return f2;
+    return function () {
+        if (f1.apply(this, arguments) === false) return false;
+        return f2.apply(this, arguments);
+    };
+}
+
+export function reverseStoppableEventChain(f1, f2) {
+    if (f1 === nop) return f2;
+    return function () {
+        if (f2.apply(this, arguments) === false) return false;
+        return f1.apply(this, arguments);
+    };
+}
+
+export function nonStoppableEventChain(f1, f2) {
+    if (f1 === nop) return f2;
+    return function () {
+        f1.apply(this, arguments);
+        f2.apply(this, arguments);
+    };
+}
+
+export function promisableChain(f1, f2) {
+    if (f1 === nop) return f2;
+    return function () {
+        var res = f1.apply(this, arguments);
+        if (res && typeof res.then === 'function') {
+            var thiz = this, args = arguments;
+            return res.then(function () {
+                return f2.apply(thiz, args);
+            });
+        }
+        return f2.apply(this, arguments);
+    };
+}
